+++ src/server/middleware/auth.js
@@ -1,0 +1,4 @@
+/**
+ * @param {import('../').default} server
+ */
+export default (server) => (req, res, next) => server.auth(req.session, next);