+++ src/server/middleware/index.js
@@ -3,6 +3,7 @@ import SessionMiddleware from './session';
 import RenderMiddleware from './render';
 import PageNotFoundMiddleware from './page-not-found';
 import ErrorMiddleware from './error';
+import AuthMiddleware from './auth';
 
 export {
 	// eslint-disable-next-line import/prefer-default-export
@@ -11,4 +12,5 @@ export {
 	RenderMiddleware,
 	PageNotFoundMiddleware,
 	ErrorMiddleware,
+	AuthMiddleware,
 };