+++ src/server/plugin.js
@@ -3,4 +3,43 @@ export default class Plugin {
 	getEntryInjections() {
 		return [];
 	}
+
+	getSocketEvents() {
+		return [];
+	}
+
+	getSocketClasses() {
+		return [];
+	}
+
+	getRouteCallbacks() {
+		return [];
+	}
+
+	getBeforeExecutions() {
+		return [];
+	}
+
+	getScripts() {
+		return [];
+	}
+
+	getStyles() {
+		return [];
+	}
+
+	getMergeStyles() {
+		return [];
+	}
+
+	getMiddlewares() {
+		return [];
+	}
+
+	// TODO register routes -> pages on client side
+	// TODO register components -> components on client side
+
+	// TODO custom layout for plugin?
+
+	// TODO plugins in rsconfig
 }